float ---> int

+++ src/BonusMilesService.java
@@ -1,7 +1,7 @@
 import java.util.Scanner;
 
 public class BonusMilesService {
-    public float calculate (float price)
+    public float calculate (int price)
     {
         return price / 20;
     }
